# specs/001-github-project-chat/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: GitHub Projects Chat API
  description: |
    REST API for the GitHub Projects Chat Interface.
    Enables natural language task management for GitHub Projects V2.
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8000/api/v1
    description: Development server
  - url: https://api.example.com/api/v1
    description: Production server
...
paths:
...
  /chat/messages:
    get:
      tags: [chat]
      summary: Get chat history
      description: Returns recent chat messages for the session
      operationId: getChatHistory
      security:
        - sessionCookie: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
      responses:
        '200':
          description: Chat history
          content:
            application/json:
              schema:
                type: object
                properties:
                  messages:
                    type: array
                    items:
                      $ref: '#/components/schemas/ChatMessage'

    post:
      tags: [chat]
      summary: Send chat message
      description: Sends user message and receives AI response
      operationId: sendChatMessage
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                  maxLength: 10000
                  example: Create a task to implement user authentication
      responses:
        '200':
          description: AI response with optional task proposal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /chat/proposals/{proposalId}/confirm:
    post:
      tags: [chat]
      summary: Confirm task proposal
      description: Confirms an AI-generated task proposal for creation
      operationId: confirmProposal
      security:
        - sessionCookie: []
      parameters:
        - name: proposalId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                edited_title:
                  type: string
                  maxLength: 256
                edited_description:
                  type: string
                  maxLength: 65535
      responses:
        '201':
          description: Task created from proposal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '410':
          description: Proposal expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session_id
  schemas:
...
    Task:
      type: object
      properties:
        task_id:
          type: string
          format: uuid
        project_id:
          type: string
        github_item_id:
          type: string
        title:
          type: string
          maxLength: 256
        description:
          type: string
          nullable: true
          maxLength: 65535
        status:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ChatMessage:
      type: object
      properties:
        message_id:
          type: string
          format: uuid
        sender_type:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
        action_type:
          type: string
          enum: [task_create, status_update, project_select]
          nullable: true
        action_data:
          type: object
          nullable: true
        timestamp:
          type: string
          format: date-time
    ChatResponse:
      type: object
      properties:
        message:
          $ref: '#/components/schemas/ChatMessage'
        proposal:
          $ref: '#/components/schemas/TaskProposal'
          nullable: true
    TaskProposal:
      type: object
      properties:
        proposal_id:
          type: string
          format: uuid
        original_input:
          type: string
        proposed_title:
          type: string
        proposed_description:
          type: string
        status:
          type: string
          enum: [pending, confirmed, edited, cancelled]
        expires_at:
          type: string
          format: date-time
    Error:
      type: object
      properties:
        error:
          type: string
        detail:
          type: string
        code:
          type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Bad Request
            detail: Invalid project_id format
            code: INVALID_PROJECT_ID
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
            detail: Session expired or invalid
            code: UNAUTHORIZED
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Not Found
            detail: Task not found
            code: NOT_FOUND